requirements.txt
flask==2.0.1
flask-cors==3.0.10
opencv-python-headless==4.5.3.56
torch==2.1.0
ultralytics==8.0.196
gunicorn==20.1.0
python-dotenv==0.19.0